Split out transform processing into per-target processors
